requirements.txt
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.12.2
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.12.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
prompt-toolkit==3.0.36
Pygments==2.18.0
python-dotenv==1.0.1
questionary==2.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
